+++ yesod-core/test/Hierarchy/Admin.hs
@@ -10,6 +10,17 @@ import Hierarchy.ResourceTree
 import Language.Haskell.TH
 import Data.Text (Text)
 
-mkRenderRouteInstanceOpts (setFocusOnNestedRoute (Just "AdminR") defaultOpts) [] (ConT ''Hierarchy) (map (fmap parseType) hierarchyResources)
-mkRouteAttrsInstanceFor [] (ConT ''AdminR) "AdminR" $ map (fmap parseType) hierarchyResources
-mkParseRouteInstanceFor "AdminR" $ map (fmap parseType) hierarchyResources
+mkRenderRouteInstanceOpts
+    (setFocusOnNestedRoute (Just "AdminR") defaultOpts)
+    []
+    []
+    (ConT ''Hierarchy)
+    (map (fmap parseType) hierarchyResources)
+mkRouteAttrsInstanceFor
+    []
+    (ConT ''AdminR)
+    "AdminR"
+    $ map (fmap parseType) hierarchyResources
+mkParseRouteInstanceFor
+    "AdminR"
+    $ map (fmap parseType) hierarchyResources