+++ yesod-core/test/Hierarchy/Nest2/NestInner.hs
@@ -8,6 +8,11 @@ import Yesod.Routes.TH
 import Hierarchy.ResourceTree
 import Language.Haskell.TH
 
-mkRenderRouteInstanceOpts (setFocusOnNestedRoute (Just "NestInner") defaultOpts) [] (ConT ''Hierarchy) (map (fmap parseType) hierarchyResources)
+mkRenderRouteInstanceOpts
+    (setFocusOnNestedRoute (Just "NestInner") defaultOpts)
+    []
+    []
+    (ConT ''Hierarchy)
+    (map (fmap parseType) hierarchyResources)
 mkRouteAttrsInstanceFor [] (ConT ''NestInner) "NestInner" $ map (fmap parseType) hierarchyResources
 mkParseRouteInstanceFor "NestInner" $ map (fmap parseType) hierarchyResources